Revert "Trigger macro syntax check after typing (#394)"
This reverts commit 97ff743c9915c1824278be2f8b15416783e1784f.
