+++ inputremapper/gui/user_interface.py
@@ -69,7 +69,6 @@ from inputremapper.gui.utils import (
     CTX_APPLY,
     CTX_WARNING,
     gtk_iteration,
-    debounce,
 )
 
 
@@ -201,9 +200,6 @@ class UserInterface:
         # hide everything until stuff is populated
         self.get("vertical-wrapper").set_opacity(0)
         self.window = window
-
-        source_view = self.get("code_editor")
-        source_view.get_buffer().connect("changed", self.check_on_typing)
 
         # if any of the next steps take a bit to complete, have the window
         # already visible (without content) to make it look more responsive.
@@ -469,16 +465,8 @@ class UserInterface:
             status_bar.push(context_id, message)
             status_bar.set_tooltip_text(tooltip)
 
-    @debounce(500)
-    def check_on_typing(self, *_):
-        """To save latest input from code editor and call syntax check."""
-        print("typing")
-        self.editor.gather_changes_and_save()
-        self.check_macro_syntax()
-
     def check_macro_syntax(self):
         """Check if the programmed macros are allright."""
-        print("check_macro_syntax")
         self.show_status(CTX_MAPPING, None)
         for key, output in active_preset:
             output = output[0]
